backend/requirements.txt
flask
flask_cors
pillow
opencv-python
yoloface
numpy
tensorflow